Added ctrl-F9, copy current line to command line.

+++ src/cpp/Edit/Xcb/EdTinp.cpp
@@ -16,7 +16,7 @@
 //       Editor: Implement EdTerm.h keyboard and mouse event handlers.
 //
 // Last change date-
-//       2023/09/19
+//       2023/09/26
 //
 //----------------------------------------------------------------------------
 #include <string>                   // For std::string
@@ -688,11 +688,15 @@ void
          activate(file);
        break;
      }
-     case XK_F9: {                  // Copy cursor line to command line
-       // (This sequence DOES NOT change the cursor line)
-       Active& active= data->active; // The current command line
-       const char* command= active.truncate(); // Truncate it
-       editor::hist->activate(command); // Activate the history view
+     case XK_F9: {                  // Copy to command line
+       if( state & gui::KS_CTRL ) { // Copy cursor line to command line
+         // (This sequence DOES NOT change the cursor line)
+         Active& active= data->active; // The current command line
+         const char* command= active.truncate(); // Truncate it
+         editor::hist->activate(command); // Activate the history view
+       } else {                     // Copy file name to command line
+         editor::hist->activate(editor::file->name.c_str());
+       }
        break;
      }
      case XK_F10: {                 // Line to top
